MINOR Add vote check, code state and toopp function
Summary

* change "my_votes" to "vote_list"
* add "STATE" global variable
* add toopp function
* add a vote check option in DEV state

The code currently supports three types of states: development (DEV),
debug (DEBUG) and production (PROD). To change this state, simply
change the value of "STATE" global variable.

In this commit, in DEV state, the code outputs the selected vote in
"vote_list" global variable. It is useful to check if the selected vote
is the desired vote for a given poll.

Note that these variable names can change. Like "my_votes" changed to
"vote_list".

The "toopp" function sums every option of every previously considered
poll to calculate the index of the options of current poll. This is
necessary because all the option of every and each poll are packed
together.