MAJOR Release 1.0.0
This is a major release. The first major release!

Summary

* add gfrestart.js with code to restart a form whenever possible
* add click to vote the desired options in a given poll
* add click to send the form
* both added clicks works only if in PROD mode
* update README.md file

The code supports now supports PROD actions: click to vote and to send
the form of the polling. It will print out the vote if in DEV state.

To change the state, simply change "STATE" variable.

It is worth noting that despite this script working, it is just a
simple script. It needs to be manually operated - which isn't that bad.

This code can be used within a browser console or with automators. I do
recomend Greasemonkey for Firefox if you want to autovote. It's the
extension used to test this code.

ONLY WORKS if the Google Form has only polls with a single selection.
Any poll with more than one option to be selected won't work.

The README.md file was updated to reflect the usage of the code in its
current form. It will only be updated if the code changes.