MINOR Add vote count cookie
Summary

* add code to count the number of votes
* store vote count in cookies
* print previous vote count in polling page
* print updated vote count in "vote completed" page

The code now counts the amount of votes. It is stored in a cookie.

The user can chose the expire date for this cookie. The expire date is
calculated in days (from now to "number of chosen days").

This code is divided in both gfvote.js (that show current number of
votes) and gfrestart.js (that shows updated vote count and updates the
number of votes). So STATE variable was also added to gfrestart.js.

It is worth noting that this function might have caused reduction of
votes per minute. It is being evauated now and future updates will
either fix it, give the option to chose to turn it off, leave it the
way it is or remove completely the function.

The README.md file was updated to reflect the usage of the code in its
current form. It will only be updated if the code changes.