MINOR Add extension-specific code and donations
Summary

* add code specific to greasemonkey
* add code specific to tampermonkey
* add donation button to README.md (PayPal and Bitcoin)
* Some minor fixes not worth mentioning

Since Greasemonkey and Tampermonkey are the default tools for
automation of code that deals with original pages, I added code for
these two extensions. Firefox uses Greasemonkey - although there is
Tampermonkey for it. Google Chrome, Chromium, Microsoft Edge and even
Safari uses Tampermonkey. These are located inside `auto` directory.

Code for manual use is still available in the `manual` directory. It's
useful for 'one time use' or just a few times use.

There is also useful metadata inserted into these extensions code. They
differ a little but it will be useful to autoupdate, for example.

The oldest extension supports less metadata but ignores metadata it
doesn't support. So it would be safe to let it be. Just in case other
extensions do not follow this behavior, I sticked to the plan of
dividing properly based on what it supports. Summing with the fact that
common users might find 'strange' not finding their extension present,
I believe I did right.

I might add a directory called `other` in the future with a script full
of metadata just in case someone wants to adapt to another extension.
But I'm not sure if I will group Greasemonkey with Tampermonkey. Let's
wait and see!

The README.md file was updated to reflect the usage of the code in its
current form. It will only be updated if the code changes. Also, a
donation button was added (PayPal and Bitcoin).