MINOR Add visual couting inside the form
Summary

* Removes console printing of the number of votes
* Adds printing of the number of votes near the form (HTML)

The output with the total number of votes was scrapped in favor of an
injection of HTML code to show the total number of votes. So, now, the
number of votes is shown easily to anyone using it.

The code was rearranjed in which `gfvote.js` and `gfrestart` do not hold
anything regarding cookies. Now that role was given to `gfcount.js` that
has is exclusively used to count votes, show on screen and store it in
cookies.

The README.md file was updated to reflect the usage of the code in its
current form. It will only be updated if the code changes.