PATCH Fix position of vote count holder
The vote count holder in mobile (small screens) was being pushed behind
the polls. It made the entire thing inutile as no vote was being shown.
It was requiring the user to roll the mouse wheel to see the vote.

A simple `z-index: 2;` fixed it. Not a big deal.

I decided to only change the version of everything IF a MINOR or MAJOR
change happens. So only the file changed itself will be holding a change
in version as it was only a PATCH.