MAJOR Code rewritten to be a single file
Summary

* made gfvote.js in the root directory
* inserted other files code into it (with proper modifications)
* deleted all directories and their files

This update makes the entire project a single file. This files supports
both Greasemonkey and Tampermonkey by default and should be runnable in
other types of script managers that uses userscripts. Maybe with a few
modifications.

Greasemonkey ignores metadata it doesn't have. So, any metadata that is
required for Tampermonley that it does not support should be ignored.

The code was properly modified to have the three files (3 version each)
into a single file without cumbersome code. I hope.

The README.md file was updated to reflect the usage of the code in its
current form. It will only be updated if the code changes.