PATCH Modify update and download URL
In the last update, the URLs were not patched after being copied. This
update makes the proper correction to those URL inside the metadata.

The README.md file was updated to reflect the usage of the code in its
current form. It will only be updated if the code changes.